requirements/base.txt
amqp==1.4.9
#anyjson==0.3.3
attrs==16.1.0
billiard==3.3.0.23
celery==3.1.24
cffi==1.8.2
click==6.6
cryptography==1.5
cssselect==0.9.2
enum34==1.1.6
Flask==0.11.1
Flask-Celery-Helper==1.1.0
Flask-Redis==0.3.0
idna==2.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.36
lxml==3.6.4
MarkupSafe==0.23
parsel==1.0.3
Pillow==3.3.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pyOpenSSL==16.1.0
pytesseract==0.1.6
pytz==2016.6.1
queuelib==1.4.2
redis==2.10.5
requests==2.11.1
Scrapy==1.1.2
scrapy-redis==0.6.3
service-identity==16.0.0
six==1.10.0
Twisted==16.4.1
w3lib==1.15.0
Werkzeug==0.11.11
zope.interface==4.3.2
-e ./undemolition
